services/infra-tenant/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.4.2
pydantic-settings==2.0.3
alembic==1.12.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
email-validator==2.1.0
